# Lammah-game-backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/questions": {
      "post": {
        "operationId": "questionsCreate",
        "summary": "Create a new question",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              },
              "examples": {
                "default": {
                  "summary": "Create approved free-game question",
                  "value": {
                    "category": "66b8f5f2c9d7a8b1e3f00222",
                    "question": "What planet is known as the Red Planet?",
                    "answer": "Mars",
                    "explanation": "Mars appears red because of iron oxide on its surface.",
                    "difficulty": "easy",
                    "points": 200,
                    "type": "text",
                    "status": "approved",
                    "source": "manual",
                    "isFreeGameQuestion": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Question created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": [
          "Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "questionsList",
        "summary": "Get all questions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Questions retrieved successfully. Answers are hidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Questions"
        ]
      }
    },
    "/questions/{id}": {
      "get": {
        "operationId": "questionsGetById",
        "summary": "Get a specific question by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Question MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00333",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question retrieved successfully. Answer is hidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Question not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Question with ID \"66b8f5f2c9d7a8b1e3f00333\" not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Questions"
        ]
      },
      "patch": {
        "operationId": "questionsUpdate",
        "summary": "Update a question",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Question MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00333",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              },
              "examples": {
                "default": {
                  "summary": "Approve and mark as free-game question",
                  "value": {
                    "status": "approved",
                    "isFreeGameQuestion": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionMutationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Question not found"
          }
        },
        "tags": [
          "Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "questionsDelete",
        "summary": "Delete a question",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Question MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00333",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Question deleted successfully"
          },
          "404": {
            "description": "Question not found"
          }
        },
        "tags": [
          "Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions": {
      "get": {
        "operationId": "adminQuestionsList",
        "summary": "Admin: get all questions including answers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Questions retrieved successfully with answers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions/ai-generated/list": {
      "get": {
        "operationId": "questionsListAiGenerated",
        "summary": "List AI-generated questions for review",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gameMode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "catalog",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions/bulk-action": {
      "post": {
        "operationId": "questionsBulkAction",
        "summary": "Apply an admin review action to questions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkQuestionActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkQuestionActionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions/{id}/retry-primary-asset": {
      "post": {
        "operationId": "questionsRetryPrimaryAsset",
        "summary": "Retry primary asset resolution",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions/{id}/retry-cover-image": {
      "post": {
        "operationId": "questionsRetryCoverImage",
        "summary": "Retry cover image resolution",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions/{id}": {
      "get": {
        "operationId": "adminQuestionsGetById",
        "summary": "Admin: get a specific question including answer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Question MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00333",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question retrieved successfully with answer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Lammah Game API",
    "description": "Runtime HTTP contract for Lammah game administration, content, and gameplay.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateQuestionDto": {
        "type": "object",
        "properties": {}
      },
      "DifficultyLevel": {
        "type": "string",
        "enum": [
          "easy",
          "medium",
          "hard"
        ]
      },
      "GameMode": {
        "type": "string",
        "enum": [
          "trivia",
          "identifyCharacter",
          "identifyVoice",
          "identifyImage",
          "completeQuote",
          "timeline",
          "emojiPuzzle",
          "identifySong",
          "identifySinger",
          "identifyMusicIntro"
        ]
      },
      "QuestionStatus": {
        "type": "string",
        "enum": [
          "draft",
          "approved",
          "published",
          "archived",
          "rejected"
        ]
      },
      "QuestionSource": {
        "type": "string",
        "enum": [
          "manual",
          "ai",
          "imported",
          "music"
        ]
      },
      "QuestionAssetType": {
        "type": "string",
        "enum": [
          "audio",
          "image",
          "video",
          "gif"
        ]
      },
      "QuestionAssetResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/QuestionAssetType"
          },
          "url": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string"
          },
          "searchQuery": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "type",
          "url",
          "source"
        ]
      },
      "AssetStatus": {
        "type": "string",
        "enum": [
          "NOT_REQUIRED",
          "PENDING",
          "READY",
          "FAILED"
        ]
      },
      "QuestionResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          },
          "correctAnswer": {
            "type": "string"
          },
          "wrongAnswers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "explanation": {
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyLevel"
          },
          "points": {
            "type": "number",
            "enum": [
              200,
              400,
              600
            ]
          },
          "score": {
            "type": "number",
            "enum": [
              200,
              400,
              600
            ]
          },
          "gameMode": {
            "$ref": "#/components/schemas/GameMode"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "source": {
            "$ref": "#/components/schemas/QuestionSource"
          },
          "primaryAsset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionAssetResponseDto"
              }
            ]
          },
          "coverImage": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionAssetResponseDto"
              }
            ]
          },
          "primaryAssetRequest": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "coverImageRequest": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "coverImageStatus": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "coverImageFailureReason": {
            "type": "string"
          },
          "assetStatus": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "assetFailureReason": {
            "type": "string"
          },
          "assetFailureStep": {
            "type": "string"
          },
          "assetFailureDiagnostics": {
            "type": "object",
            "additionalProperties": true
          },
          "gameplayMetadata": {
            "type": "object",
            "additionalProperties": true
          },
          "aiMetadata": {
            "type": "object",
            "additionalProperties": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "mediaUrl": {
            "type": "string"
          },
          "hasPreviewAudio": {
            "type": "boolean"
          },
          "isFreeGameQuestion": {
            "type": "boolean"
          },
          "qualityScore": {
            "type": "number"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "_id",
          "question",
          "wrongAnswers",
          "difficulty",
          "status",
          "source"
        ]
      },
      "QuestionMutationResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/QuestionResponseDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "data",
          "message"
        ]
      },
      "QuestionListResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponseDto"
            }
          }
        },
        "required": [
          "statusCode",
          "data"
        ]
      },
      "QuestionDetailResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/QuestionResponseDto"
          }
        },
        "required": [
          "statusCode",
          "data"
        ]
      },
      "UpdateQuestionDto": {
        "type": "object",
        "properties": {}
      },
      "BulkQuestionActionDto": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "action": {
            "type": "string",
            "enum": [
              "approve",
              "reject",
              "delete"
            ]
          }
        },
        "required": [
          "ids",
          "action"
        ]
      },
      "BulkQuestionActionResponseDto": {
        "type": "object",
        "properties": {
          "modifiedCount": {
            "type": "number"
          }
        },
        "required": [
          "modifiedCount"
        ]
      },
...
    }
  }
}