# Lammah-game-frontend/src/api/generated/models/updateQuestionDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lammah Game API
 * Runtime HTTP contract for Lammah game administration, content, and gameplay.
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateQuestionDto {
  [key: string]: unknown;
}